create polkadot api once before processing blocks
