+++ squid-blockexplorer/src/processor.ts
@@ -6,6 +6,7 @@ import {
     BatchProcessorItem,
 } from '@subsquid/substrate-processor';
 import { Store, TypeormDatabase } from '@subsquid/typeorm-store';
+import { ApiPromise, WsProvider } from "@polkadot/api";
 import config from './config';
 import { processBalancesFactory, createProcessBalancesDependencies } from './balances';
 import { processBlocksFactory, createProcessBlocksDependencies } from './blocks';
@@ -29,7 +30,20 @@ const processor = new SubstrateBatchProcessor()
 processor.run(new TypeormDatabase(), processChain);
 
 async function processChain(ctx: Context): Promise<void> {
-    const processBlocksDependencies = await createProcessBlocksDependencies(ctx);
+    const types = {
+        Solution: {
+            public_key: 'AccountId32',
+            reward_address: 'AccountId32',
+        },
+        SubPreDigest: {
+            slot: 'u64',
+            solution: 'Solution',
+        }
+    };
+
+    const provider = new WsProvider(process.env.CHAIN_RPC_ENDPOINT);
+    const api = await ApiPromise.create({ provider, types });
+    const processBlocksDependencies = createProcessBlocksDependencies(ctx, api);
     const processBalancesDependencies = createProcessBalancesDependencies(ctx);
     const processBlocks = processBlocksFactory(processBlocksDependencies);
     const processBalances = processBalancesFactory(processBalancesDependencies);
